requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
django-cors-headers
djangorestframework-jwt
mysqlclient
python-dotenv